energy_forecasting_project/requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0,<0.13.0

# Machine learning libraries
scikit-learn>=1.1.0
tensorflow-gpu>=2.10.0,<2.16.0
xgboost>=1.6.0

# Python version requirement
python>=3.10,<3.12

# Web framework
flask>=2.2.0
fastapi>=0.85.0
uvicorn>=0.18.0

# Streamlit for interactive dashboards
streamlit>=1.12.0

# Additional utilities
jupyter>=1.0.0
plotly>=5.10.0
joblib>=1.2.0
